+++ examples/demons_registration_2d.py
@@ -37,13 +37,13 @@ def main():
 
     # In order to use a GPU uncomment the following line. The number is the device index of the used GPU
     # Here, the GPU with the index 0 is used.
-    device = th.device("cuda:0")
+    # device = th.device("cuda:0")
 
     # create test image data
     fixed_image, moving_image, shaded_image = create_C_2_O_test_images(256, dtype=dtype, device=device)
 
     # create pairwise registration object
-    registration = al.DemonsRegistraion(verbose=False)
+    registration = al.DemonsRegistraion(verbose=True)
 
     # choose the affine transformation model
     transformation = al.transformation.pairwise.NonParametricTransformation(moving_image.size,